Fix #297 adding deflate and gzip handling
Tested on py24, py27 and py32 and the failing tests on 3.2 go away, the content
we're getting back is still gzipped so this is a valid manual test.  Add
BytesIO to backwardcompat, fallingback to StringIO on py2x.

I'd like to get the CI build green then comeback and improve the testing around
this.  I've not added tests at the moment, but want to come back to that. There is
quite a lot of complexity around downloads split between pip.index and
pip.download.