Fix for subtle config override bug
Config keys should be normalized (i.e., have underscores replaced with
dashes) *before* the config dictionaries are merged.

Otherwise, if you set a value in a config file and via an environment
variable both the dash separated key (from the config file) and the
underscore separated key (from the environment variable) end up in the
config dictionary. Usually this doesn't matter because the latter key
comes after the former when the dictionary is iterated over. But in
cases where the two config keys hash to the same value module the size
of the dictionary, the order is reversed and the wrong value takes
precedance.

For instance with 'index-url' there's no problem:

    >>> repr({'index-url': 1, 'index_url': 2})
    "{'index-url': 1, 'index_url': 2}"

But with 'no-index', the value from the config file overwrites the
value from the environment variable:

    >>> repr({'no-index': 1, 'no_index': 2})
    "{'no_index': 2, 'no-index': 1}"