remove basecommand.merge_options()
This patch removes the error prone step of merging the baseparsers'
options (the 'Generic Options') with those of the subparser. In fact,
commands are already capable of parsing all options present in the
'Generic Options' option group. The option parsing logic is as follows:

Given a command line such as:
  '--timeout 5 install --user pkg'
pip.parseopt() returns command 'install' with arguments:
  '--timeout 5 --user pkg'
pip.main() executes the above as:
  commands['install'](parser).main('--timeout 5 --user pkg', options)