Update installing.rst
These are the appropriate instructions, as I best understand them, for installing pip and virtualenv globally on Linux, Mac OS X and Windows. I never knew an OS could make it so difficult to save a file with the name you want...

Even if we do a "get-virtualenv.py" script instead (which is a good idea), the challenges of downloading Python scripts with IE and ensuring they have the right name will still apply. I think we also want to unconditionally recommend installing the Python Launcher for anyone that is using 2.7 on Windows rather than 3.3+