Fix for unicode in the log on Python 2
This won't work on Python 3, and is probably not required there at all.

It makes sure that all entries from complete_log are bytestrings when
join()ing, before writing to the file.

The issue (github #1137) was that a single unicode object on the list
would cause all the bytestrings to be converted to unicode, probably
through the 'ascii' encoder (which would then fail on 8-bit characters).