Condense --allow-external and --allow-unverified into one option.
The decision to support --allow-external and --allow-unverified was
made in order to allow people to safely host their files externally
to the index (generally PyPI). However since the implementation of
this it has become abundantly clear that people are simply not
choosing to host their files externally but safely. Additionally
the difference between a external but safe and an external but unsafe
file is dificult to explain and leaves users confused.

This will condense the two into a single option, --allow-external
which has the semantics of both options specified together. This
means that it is no longer possible to select safely externally
hosted files but exclude unsafely externally hosted files.