Fix broken progress bar on Windows
Checks for log level for progress bar fail when terminal's stdout
is wrapped with colorama on Windows. The proper fix is to check for
if consumer is terminal/console (isatty) rather than it is stdout.

But it looks like proper fix would break code in util.py, so this
should do the trick in short term.