Drop PyPy from the de-bundling tests
PyPy isn't often packages by downstream, and even when it it is
it's often not got a pip package to go with it, and if it does
then it's probably not very likely that this code will fail on
PyPy but not on CPython 2.7.

So we'll just drop them so that the tests complete faster.