More robust passing of egg_info --egg-base argument
A custom package setup.py can be arranged in a way that it changes
current working directory on its own (perhaps because pip doesn't
currently support specifying subdirectory for non-editable installs).
As a consequence, relative path of directory passed as --egg-base
argument of "setup.py egg_info" invocation when installing such
a package could result in a failure.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>