Make --require-virtualenv output more info
if not in a virtualenv **and** --user wasn't specified.

Previously, this would print an error even if --user was specified.
I believe the spirit of `--require-virtualenv` is that it was meant to
be used as `require_virtualenv = True` in the pip config file and what
it really seeks to do is to prevent the user from messing with packages
in the system python envvironment. Thus, I'd argue that --user shouldn't
cause an error, because it doesn't mess with the system python.

With this change, output looks like this:

    $ python -m pip install --require-virtualenv foo
    ******************************************************
    Using pip outside a virtualenv and without --user
    is not recommended!

    Doing stuff with the system python's packages can break
    your system.

    Did you mean to use a virtualenv or use --user?
    ******************************************************