tox.ini: Pin pep8<1.6.2 in main pep8 tox targets
Because pep8 keeps adding new checks and causing Travis CI to fail.

I also added new tox targets that *don't* pin pep8 and these are
reference in .travis.yml but are allowed to fail (see
http://docs.travis-ci.com/user/build-configuration/#Rows-That-are-Allowed-To-Fail).

This way we know when there are new pep8 versions with new checks that
we fail, *but* we don't unfairly penalize ourselves.