Don't bundle certifi
The idea behind requests supporting the use of certifi is so that
an installation can get newer certificates simply by updating their
copy of certifi instead of requiring a new version of requests. By
bundling certifi we're essentially reverting this behavior, forcing
people to update their copy of pip to get a new copy of certifi to
update their certififcaates. Instead we'll use a globally installed
certifi if it's available which will allow people to upgrade their
certificates as requests intended.