tox.ini: Add py34fast target
This runs just the unit tests which don't need the network, so it could
be useful for doing a quick sanity check of some changes.

The regular tox targets take so long that they are not conducive to TDD
and I am often tempted to skip them and that's not good.

inspired in part by seeing how fast the unit tests were (and that you can
filter out the ones that require network access) in
https://github.com/pypa/pip/issues/2498