tox.ini: Add "cover" target
that measures coverage for unit tests only. Functional tests are slow
enough already and a lot of them fail when coverage is measured, because
they fork processes and them emit warnings to stderr that the tests are
not expecting.

I think it's a good thing to try to increase unit test coverage and this
lets us measure and see where the gaps are. And it will help folks
realize if they increased or decreased coverage with their changes.