pip.models.index: Add doctests
These are very trivial and overkill for a simple class, but I'm hoping
that it encourages adding docstrings for future additions to this
module. I also got confused myself about what `url_to_path` does and
I think I wrote that myself not too long ago, so docs are a good thing.
:)