Fix only-download use in req_set.
only_download means 'mark the extracted tree for automatic deletion'.
is_download in req_set indicates we're downloading to a user specified
directory - and has nothing to do with deleting the extracted tree.
The actual thing we want to use to control that is the removed
--no-install parameter.

Additionally for wheels we only want to mark the unpacked tree for
deletion when we're downloading the wheel (rather than installing it)
as otherwise the marker looks like part of the wheel.