Issue #2839: Revert syntax of requirement files.
In < 7 requirements files did not need escaping for quotation marks.
7.0 accidentally introduced quoting (via the use of shlex.split) when
also introducing the feature of supporting local and global options.
Since we've only had a week while this has been public it should be
pretty safe to unwind the quoting aspect and restore previously
working requirements files to working order.