Add two new options to the pip install command that are meant to give the user more control when downloading packages.
The two new options are --platform and --supported-interpreter-version, and both enforce that the --download option is also specified, else a CommandError is raised.

With the --platform option, a user can ask to download wheels of a different platform than that of the local machine running the command, which is designed as a utility for gathering dependencies and preparing a distribution. Because this option enforces that the --download option is also specified, it will never attempt to install wheels of an incorrect platform.

With the --supported-interpreter-version option, a user can ask to download wheels that are explicitly compatible with a specific Python interpreter version, which is designed as a similar utility to the --package option.