Add --platform and --python-version to `pip download`.
With the --platform option, a user can download wheels with
a different platform than that of the local machine running the command.

With the --python-version option, a user can
download wheels that are explicitly compatible with a specific
Python interpreter version.

This functionality is meant for utilities that gather dependencies
and prepare distributions for other platforms.