Fix for 3141: symlink in consolescript path
This fixes issue 3141, where a uninstall error occurs
when upgrading/uninstalling and the path to the console
script contains symlinks. As the path is in installed-files.txt
as well and these paths are joined to the real path coming
from .egg_info, the uninstall will try to remove the
console script using two different paths and fail on the second
(non-real) path.