Drop markers inside InstallRequirement.check_if_exists
Any requirement that are checked if they are installed may specify
an environment marker. Since the marker may include information that
we no longer have context for (such as which extra requires that
requirement), and we have already decided that we are going to install
the requirement, drop the marker when we check if it's installed.