Warn when installing scripts that will not be on PATH
It is common for systems with minimal shell configuration to exclude Python's executable paths from PATH (e.g. OS X, or --user on other systems).
This can be confusing, because no mention is made of installed scripts,
or their destination.

This adds a warning indicating when installed scripts are not on PATH,
including exactly where they are.