Make corrections in new tests
The tests added for non-eager upgrading checks were passing but they
were not checking if things got installed. Upon addition of these
checks, the tests started failing sincethey were not using the
data/packages directory.

This patch fixes the tests to run pip with correct arguments and check
for the right things.