Draft implementation of pep-503 data-requires-python
This allows pip to understand the `data-requires-python` metadata
information that can be set on a simple repository. This allows pip to
ignore any release or file that would not be compatible with the current
Python version even before trying to download and install this version.

Relevant extract of pep 503 at the time of this writing.

    A repository MAY include a data-requires-python attribute on a file
    link. This exposes the Requires-Python metadata field, specified in PEP
    345 , for the corresponding release. Where this is present, installer
    tools SHOULD ignore the download when installing to a Python version
    that doesn't satisfy the requirement. For example:

    <a href="..." data-requires-python="&gt;=3">...</a>
    In the attribute value, < and > have to be HTML encoded as &lt; and &gt;
    , respectively.

This can mostly be used to, for example, mark a new sdist of a new
package version as requires-python >3.4, and not  fail to install or
upgrade on users systems.

This will require extra patches to PyPI-legacy and warehouse to be
usable. Though releasing a version of pip that understand this feature
is necessary to have wide adoption at the time when these metadata get
actually published.