Fix pip when using Py2 on Windows with non-ASCII hostname or username.
Two related problems are fixed:
- Previously non-ASCII characters in hostname blew up `pip install`
  completely. This is rather severe.
- Non-ASCII characters in username crashed printing help text. Not
  so bad but definitely annoying.

Non-ASCII usernames are pretty common in non-English speaking
countries. That also makes non-ASCII hostnames pretty common, because
Windows creates hostname based on the username by default.

The reason for these failures was that `user_cache_dir` returned
Unicode on Windows and bytes elsewhere, and rest of the codebase was
expecting paths to be bytes.

It could be argued that pip should always use Unicode internally, but
that would require a lot more changes and fixing also some of the
vendored dependencies. We can also argue, like PEP-519 does, that
paths should not be considered to be strings at all, and thus the "all
Unicode internally" guideline wouldn't apply in this case.

Fixes #3463. See that issue for more discussion and details.