Support relative paths for editable file links.
When searching paths in the `InstallRequirement` class, use
`os.path.abspath(self.source_dir)` whenever referring directly to an
artifact in the filesystem. Doing so supports relative file paths
properly.

The `SETUPTOOLS_SHIM` compiles and executes a `setup.py` for an editable
file link. The form of the command opens directly the path from
`file:<path>` links while also setting `cwd=<path>`. This may break
relative paths where the dependent project is at a different depth in a
source tree from the referenced requirement. In that case, the opened
file `<path>` relative to the relative path `<path>` will not align,
leading to an `OSError`.