Fix the latest version number in search results
When listing found packages, `print_results()` uses the last version in
the versions list of a package, but that version is not necessarily the
latest. For example, `SearchCommand.search()` may return the following
(tested with JFrog Artifactory PyPI repo):

    pypi_hits = command.search(args, options)

    [{'_pypi_ordering': 1,
      'name': 'testlib',
      'summary': 'Test library.',
      'version': '1.0.5'},
     {'_pypi_ordering': 0,
      'name': 'testlib',
      'summary': 'Test library.',
      'version': '1.0.3'}]

`pip.commands.search.transform_hits()` preserves the order of versions:

    hits = pip.commands.search.transform_hits(pypi_hits)

    [{'name': 'testlib',
      'summary': 'Test library.',
      'versions': ['1.0.5', '1.0.3']}]

`pip.commands.search.print_results()` uses the last version in the list:

    pip.commands.search.print_results(hits)

    testlib (1.0.3)  - Test library.

Expected result:

    testlib (1.0.5)  - Test library.

`print_results()` uses `highest_version()` already to compare the
installed version with the latest found version (for installed
packages). Same function is used in this fix to show the latest
versions of all found packages.