workaround a Windows 'already removed' rmtree problem
On Windows, we can get an Access Denied error as a result of our rmtree()
operation, but with the original operation actually removing the target
folder. This could cause problems with the repeated operation failing due
to the target folder (or some part of it) not existing.

Workaround is to treat any 'does not exist` errors as success.