Add a new *public* main() method
This is mostly compatible with the prior use of `pip.main()` but not
exactly identical.
If the recommended mode of use for invoking `pip` programmatically is to
call `sys.executable` with `-m pip`, then that mode of use *must* be
supported by `pip`.

`subprocess.check_output(...)` behaves very similarly, but of course not
identically, to `pip.main(...)` on pip<9.0.2

This is therefore a mostly compatible interface which alleviates some
transitional hurdles for those using virtualenv or other 3rd party
packages which depend upon `pip.main`.
The only packages which will remain broken are those which use
`pip.main` from an environment in which subprocess cannot be spawned.

The new public `pip.main` is very clearly declared as the *only* public
python API for pip and is noted as having been added for better
backwards compatibility with known (but not supported) usage.