appdirs: return paths relative to effective user's homedir
Since `expanduser()` is used to expand the `~` to the home directory,
`~/.cache` will always return the home directory of the user under which
the python process is running. However, if pip is executed from within
a Python process in which the euid has changed, this means that the
cache dir (and other paths) will not be writable, causing pip to fail.

This commit fixes that issue by attempting to get the username matching
the euid, and using that username (e.g. `~user/.cache`) when invoking
`expanduser()`.